+++ api/polyfills.js
@@ -1,0 +1,1 @@
+export default async () => Response.redirect('https://unpkg.com/@shgysk8zer0/polyfills@0.4.0/all.min.js');