+++ mimes.js
@@ -1,1 +1,0 @@
-export * from '@shgysk8zer0/consts/mimes.js';