+++ tests/echo.js
@@ -1,29 +1,30 @@
 import { RequestHandlerTest } from '../RequestHandlerTest.js';
 
-const url = new URL('http://localhost:8888/api/echo');
-const origin = 'http://localhost:9999';
+const origin = 'http://localhost:8888';
+const url = new URL('/api/echo', origin);
+const referrer = origin;
 const body = new FormData();
 body.set('foo', 'bar');
 body.set('file', new File(['Hello, World!'], 'hi.txt', { type: 'text/plain' }));
 
+const headers = {
+	'Sec-Fetch-Dest': 'empty',
+	'Sec-Fetch-Mode': 'no-cors',
+	'Sec-Fetch-Site': 'same-origin',
+	'Origin': origin,
+	'Referer': referrer,
+};
+
 const { error } = await RequestHandlerTest.runTests(
 	new RequestHandlerTest(
-		new Request(url + '?1', { headers: new Headers({ Origin: origin }) }),
-		[RequestHandlerTest.shouldAllowOrigin, RequestHandlerTest.shouldBeOk]
-
+		new Request(url + '?test=basic', { headers, referrer }),
+		[RequestHandlerTest.shouldRequireSameOrigin, RequestHandlerTest.shouldBeOk, RequestHandlerTest.shouldBeJSON]
 	),
 	new RequestHandlerTest(
-		new Request(url + '?2', {
-			method: 'PUT',
-			headers: new Headers({ Origin: origin }),
-			body: body.get('file'),
-		}),
-		RequestHandlerTest.shouldBeOk
-	),
-	new RequestHandlerTest(
-		new Request(url + '?3', {
-			method: 'FOO',
-			headers: new Headers({ Origin: origin }),
+		new Request(url + '?test=invalid=method', {
+			method: 'PATCH',
+			headers,
+			body,
 		}),
 		RequestHandlerTest.shouldNotAllowMethod
 	),
@@ -32,8 +33,27 @@ const { error } = await RequestHandlerTest.runTests(
 		RequestHandlerTest.shouldClientError
 	),
 	new RequestHandlerTest(
-		new Request(url + '?5', { headers: { Origin: 'https://not-allowed.org' }}),
-		RequestHandlerTest.shouldDisallowOrigin
+		new Request(url + '?test=cross-origin', {
+			headers: {
+				...headers,
+				Origin: 'https://not-allowed.org',
+				Referer: 'about:client'
+			},
+			mode: 'cors',
+			referrerPolicy: 'no-referrer',
+		}),
+		[RequestHandlerTest.shouldDisallowOrigin, RequestHandlerTest.shouldRequireSameOrigin, RequestHandlerTest.shouldClientError]
+	),
+	new RequestHandlerTest(
+		new Request(url + '?test=options', {
+			method: 'OPTIONS',
+			referrer,
+			headers: {
+				'Access-Control-Request-Method': 'POST',
+				...headers
+			}
+		}),
+		[RequestHandlerTest.shouldBeOk, RequestHandlerTest.shouldAllowMethod]
 	)
 );