+++ CHANGELOG.md
@@ -7,6 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v1.1.0] - 2024-08-20
+
+### Added
+- Add `NetlifyRequest`
+- Add `DocumentTemplate`
+- Add multiple new tests
+
+### Changed
+- In `createHandler`, use `NetlifyRequest` to fix errors in `Request` objects
+- Update ESLint to use `.gitignore` and new config
+
 ## [v1.0.2] - 2024-08-02
 
 ### Added